Completed entire front end

+++ app.py
@@ -14,6 +14,8 @@ app = Flask(__name__)
 app.secret_key = 'OIUHs9w8zqJSHd112' #Random string for session
 app.config['SESSION_COOKIE_NAME'] = 'Joshs Cookie'
 app.config['UPLOAD_FOLDER'] = os.path.join('static' , 'Images')
+app.config["TEMPLATES_AUTO_RELOAD"] = True
+
 TOKEN_INFO = "token_info"
 
 @app.route('/')
@@ -23,18 +25,20 @@ def index():
     #Removes images in folder
     for filename in os.listdir(app.config['UPLOAD_FOLDER']):
         file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+        if(filename[0] == 'a'): #Only deletes albums
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
 
     sp = create_spotify_oauth()
     auth_url = sp.get_authorize_url()
 
-    return redirect(auth_url)
+    return render_template("index.html", url=auth_url)
+    # return redirect(auth_url)
 
 @app.route('/redirect')
 def redirectPage():
@@ -87,6 +91,10 @@ def songs():
         
         track = sp.track(track_id=uri)
         trackName = track['name']
+        if(len(trackName) > 25):
+            trackName = trackName[:25] + "..."
+
+
         trackNames.append(trackName)
 
         imageURL = track['album']['images'][1]['url']
@@ -104,21 +112,21 @@ def songs():
         artist = artist.rsplit(',' , 1)[0]
         artists.append(artist)
         
-        # print(trackName)
+        #print(trackName)
         # print(artists)
     
     return render_template(
         "songs.html",  
-        albumImg1 = imagePaths[0], link1 = trackURLs[0], name1 = trackName[0], artist1 = artists[0],
-        albumImg2 = imagePaths[1], link2 = trackURLs[1], name2 = trackName[1], artist2 = artists[1],
-        albumImg3 = imagePaths[2], link3 = trackURLs[2], name3 = trackName[2], artist3 = artists[2],
-        albumImg4 = imagePaths[3], link4 = trackURLs[3], name4 = trackName[3], artist4 = artists[3],
-        albumImg5 = imagePaths[4], link5 = trackURLs[4], name5 = trackName[4], artist5 = artists[4],
-        albumImg6 = imagePaths[5], link6 = trackURLs[5], name6 = trackName[5], artist6 = artists[5],
-        albumImg7 = imagePaths[6], link7 = trackURLs[6], name7 = trackName[6], artist7 = artists[6],
-        albumImg8 = imagePaths[7], link8 = trackURLs[7], name8 = trackName[7], artist8 = artists[7],
-        albumImg9 = imagePaths[8], link9 = trackURLs[8], name9 = trackName[8], artist9 = artists[8],
-        albumImg10= imagePaths[9], link10 = trackURLs[9],name10= trackName[9], artist10 = artists[9],
+        albumImg1 = imagePaths[0], link1 = trackURLs[0], name1 = trackNames[0], artist1 = artists[0],
+        albumImg2 = imagePaths[1], link2 = trackURLs[1], name2 = trackNames[1], artist2 = artists[1],
+        albumImg3 = imagePaths[2], link3 = trackURLs[2], name3 = trackNames[2], artist3 = artists[2],
+        albumImg4 = imagePaths[3], link4 = trackURLs[3], name4 = trackNames[3], artist4 = artists[3],
+        albumImg5 = imagePaths[4], link5 = trackURLs[4], name5 = trackNames[4], artist5 = artists[4],
+        albumImg6 = imagePaths[5], link6 = trackURLs[5], name6 = trackNames[5], artist6 = artists[5],
+        albumImg7 = imagePaths[6], link7 = trackURLs[6], name7 = trackNames[6], artist7 = artists[6],
+        albumImg8 = imagePaths[7], link8 = trackURLs[7], name8 = trackNames[7], artist8 = artists[7],
+        albumImg9 = imagePaths[8], link9 = trackURLs[8], name9 = trackNames[8], artist9 = artists[8],
+        albumImg10= imagePaths[9], link10 = trackURLs[9],name10= trackNames[9], artist10 = artists[9],
     )
 
 def get_token():
